labellab_ML-Models/requirements.txt
tensorflow==1.14.0
numpy==1.18.1
matplotlib==3.1.1
tensorflow-hub==0.6.0
Pillow==6.2.1
python-dotenv==0.12.0
Flask==1.0.3
